+++ source/renderer/app/components/news/NewsItem.js
@@ -39,7 +39,8 @@ export default class NewsItem extends Component<Props, State> {
     newsItemCollapsible: true,
   };
 
-  componentWillReceiveProps(nextProps: Props) {
+  // eslint-disable-next-line
+  UNSAFE_componentWillReceiveProps(nextProps: Props) {
     const { newsItemExpanded } = this.state;
     if (
       this.props.isNewsFeedOpen &&