+++ source/renderer/app/containers/wallet/WalletUtxoPage.js
@@ -16,7 +16,8 @@ type Props = InjectedProps;
 export default class WalletSettingsPage extends Component<Props> {
   static defaultProps = { actions: null, stores: null };
 
-  componentWillMount() {
+  // eslint-disable-next-line
+  UNSAFE_componentWillMount() {
     this.props.actions.walletSettings.startWalletUtxoPolling.trigger();
   }