+++ source/renderer/app/themes/daedalus/light-blue.js
@@ -251,7 +251,8 @@ export const LIGHT_BLUE_THEME_OUTPUT = {
   },
   hardwareWallet: {
     '--theme-hardware-wallet-step-background-color': '#fafbfc',
-    '--theme-hardware-wallet-step-border-active-color': '#296fd0',
+    '--theme-hardware-wallet-step-border-active-color':
+      'rgba(94, 96, 102, 0.5)',
     '--theme-hardware-wallet-step-border-color': '#c6cdd6',
     '--theme-hardware-wallet-step-color': '#5e6066',
     '--theme-hardware-wallet-title-color': '#5e6066',
@@ -946,7 +947,6 @@ export const LIGHT_BLUE_THEME_OUTPUT = {
   walletSettings: {
     '--theme-wallet-settings-section-separator-color':
       'rgba(94, 96, 102, 0.15)',
-    '--theme-wallet-settings-resync-button-label-color': 'rgba(94, 96, 102, 1)',
   },
   walletNotRespondingOverlay: {
     '--theme-wallet-not-responding-background-color': 'rgba(36, 62, 98, 0.96)',