+++ source/renderer/app/themes/daedalus/yellow.js
@@ -241,12 +241,12 @@ export const YELLOW_THEME_OUTPUT = {
     '--font-mono': 'SFMono-Light',
   },
   hardwareWallet: {
-    '--theme-hardware-wallet-step-background-color': '#fafbfc',
-    '--theme-hardware-wallet-step-border-active-color': '#296fd0',
-    '--theme-hardware-wallet-step-border-color': '#c6cdd6',
-    '--theme-hardware-wallet-step-color': '#5e6066',
-    '--theme-hardware-wallet-title-color': '#5e6066',
-    '--theme-hardware-wallet-message-color': '#5e6066',
+    '--theme-hardware-wallet-step-background-color': '#ffffff',
+    '--theme-hardware-wallet-step-border-active-color': 'rgba(45, 45, 45, 0.5)',
+    '--theme-hardware-wallet-step-border-color': '#e1dac6',
+    '--theme-hardware-wallet-step-color': '#2d2d2d',
+    '--theme-hardware-wallet-title-color': '#2d2d2d',
+    '--theme-hardware-wallet-message-color': '#2d2d2d',
   },
   icon: {
     '--theme-icon-nav-color': 'rgba(45, 45, 45, 0.6)',
@@ -939,7 +939,6 @@ export const YELLOW_THEME_OUTPUT = {
   },
   walletSettings: {
     '--theme-wallet-settings-section-separator-color': 'rgba(45, 45, 45, 0.15)',
-    '--theme-wallet-settings-resync-button-label-color': 'rgba(45, 45, 45, 1)',
   },
   walletNotRespondingOverlay: {
     '--theme-wallet-not-responding-background-color': '#fdcd68f5',