+++ source/renderer/app/themes/utils/createTheme.js
@@ -1095,7 +1095,6 @@ export const createDaedalusComponentsTheme = (
       '--theme-wallet-settings-section-separator-color': `${chroma(
         text.primary
       ).alpha(0.1)}`,
-      '--theme-wallet-settings-resync-button-label-color': `${text.primary}`,
     },
     walletNotRespondingOverlay: {
       '--theme-wallet-not-responding-background-color': `${chroma(